Move plugin index for MeteorStubs into internal directory.
