+++ npm-packages/meteor-vite/src/bin/debug/vite-server.ts
@@ -1,7 +1,7 @@
 import FS from 'fs';
 import Path from 'path';
 import { createServer } from 'vite';
-import { MeteorStubs } from '../../plugin';
+import { MeteorStubs } from '../../plugin/internal';
 
 const meteorRoot = Path.join(process.cwd(), '../../examples/vue');
