--- npm-packages/meteor-vite/src/plugin/index.ts
@@ -1,1 +1,0 @@
-export { MeteorStubs } from './internal/MeteorStubs';
