+++ npm-packages/meteor-vite/src/plugin/internal/index.ts
@@ -1,0 +1,1 @@
+export { MeteorStubs } from './MeteorStubs';
